chore(calldata): split codec/de into multiple files ♻️
